2004-03-31 00:55 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/source/rtl/fparse.c   * xharbour/source/rtl/hbcomprs.c   * xharbour/source/rtl/hbffind.c   * xharbour/source/rtl/hbrandom.c   * xharbour/source/rtl/philes.c   * xharbour/source/vm/hash.c   * xharbour/source/vm/inet.c   * xharbour/source/vm/thread.c     * warning fixed
  * xharbour/source/rtl/filesys.c
    * fixed warnings
    * some small fixes for errno settings and cleanups

+++ xharbour/source/rtl/philes.c
@@ -1,5 +1,5 @@
 /*
- * $Id: philes.c,v 1.16 2004/02/14 01:29:42 andijahja Exp $
+ * $Id: philes.c,v 1.17 2004/03/18 03:58:37 ronpinkas Exp $
  */
 
 /*
@@ -129,7 +129,7 @@ HB_FUNC( FWRITE )
    if( ISNUM( 1 ) && ISCHAR( 2 ) )
       hb_retnl( hb_fsWriteLarge( hb_parnl( 1 ),
                                  ( BYTE * ) hb_parcx( 2 ),
-                                 ISNUM( 3 ) ? hb_parnl( 3 ) : hb_parclen( 2 ) ) );
+                                 ISNUM( 3 ) ? (ULONG) hb_parnl( 3 ) : hb_parclen( 2 ) ) );
    else
       hb_retnl( 0 );
 }
